Refactor loaders; fix #145
