+++ backend/lambda_functions/lipas_loader/base_loader.py
@@ -1,0 +1,1 @@
+../base_loader/base_loader.py
