+++ backend/lambda_functions/osm_loader/osm_loader.py
@@ -1,134 +1,17 @@
-import datetime
-import json
-import logging
-import os
-import sys
-import traceback
-from typing import Any, Dict, List, Optional, Type, TypedDict
+from typing import Any, Dict, List, Optional
 
-import boto3
 import requests
 from shapely.geometry import Point, Polygon
-from sqlalchemy import MetaData, create_engine, inspect
-from sqlalchemy.exc import SQLAlchemyError
-from sqlalchemy.ext.automap import AutomapBase, automap_base
-from sqlalchemy.orm import Session, sessionmaker
 
-KoosteBase = automap_base(metadata=(MetaData(schema="kooste")))
-
-LOGGER = logging.getLogger()
-LOGGER.setLevel(logging.INFO)
+from .base_loader import LOGGER, BaseLoader, Event, Response, base_handler
 
 
-class Event(TypedDict):
-    pages: Optional[List[int]]
-
-    # Optional values
-    do_not_update_timestamp: Optional[bool]
-
-    close_to_lon: Optional[float]
-    close_to_lat: Optional[float]
-    radius: Optional[float]
-
-
-class Response(TypedDict):
-    statusCode: int  # noqa N815
-    body: str
-
-
-class DatabaseHelper:
-    def __init__(self):
-        if os.environ.get("READ_FROM_AWS", "1") == "1":
-            session = boto3.session.Session()
-            client = session.client(
-                service_name="secretsmanager",
-                region_name=os.environ.get("AWS_REGION_NAME"),
-            )
-            self._credentials = json.loads(
-                client.get_secret_value(SecretId=os.environ.get("DB_SECRET_RW_ARN"))[
-                    "SecretString"
-                ]
-            )
-        else:
-            self._credentials = {
-                "username": os.environ.get("RW_USER"),
-                "password": os.environ.get("RW_USER"),
-            }
-
-        self._host = os.environ.get("DB_INSTANCE_ADDRESS")
-        self._db = os.environ.get("DB_MAIN_NAME")
-        self._port = os.environ.get("DB_INSTANCE_PORT", "5432")
-        self._region_name = os.environ.get("AWS_REGION_NAME")
-
-    def get_connection_parameters(self) -> Dict[str, str]:
-        return {
-            "host": self._host,
-            "port": self._port,
-            "dbname": self._db,
-            "user": self._credentials["username"],
-            "password": self._credentials["password"],
-        }
-
-    def get_connection_string(self) -> str:
-        db_params = self.get_connection_parameters()
-        return (
-            f'postgresql://{db_params["user"]}:{db_params["password"]}'
-            f'@{db_params["host"]}:{db_params["port"]}/{db_params["dbname"]}'
-        )
-
-
-class Syncher:
-    def __init__(self, prepared_base: AutomapBase, session: Session):
-        self.base = prepared_base
-        self.session = session
-        self.pk_names: Dict[str, str] = {}
-        self.existing_pks: Dict[str, set] = {}
-        self.pks_to_save: Dict[str, set] = {}
-        for klass in self.base.classes:
-            # We have to use class name as key, even though class is hashable.
-            # Seems like dynamically created sqlalchemy types may not retain
-            # their hash across calls :(
-            self.pk_names[klass.__name__] = inspect(klass).primary_key[0].name
-            existing_pk_rows = session.query(
-                getattr(klass, self.pk_names[klass.__name__])
-            ).all()
-            existing_pk_set = set([row[0] for row in existing_pk_rows])
-            self.existing_pks[klass.__name__] = existing_pk_set
-
-    def mark(self, instance: object):
-        pk = getattr(instance, self.pk_names[type(instance).__name__])
-        if type(instance).__name__ not in self.pks_to_save.keys():
-            self.pks_to_save[type(instance).__name__] = set()
-        self.pks_to_save[type(instance).__name__].add(pk)
-
-    def finish(self, session: Session) -> int:
-        # We only ever delete those tables that are marked by the loader
-        pks_to_delete = {
-            name: self.existing_pks[name] - self.pks_to_save[name]
-            for name in self.pks_to_save.keys()
-        }
-        deleted = 0
-        for klass in self.base.classes:
-            if klass.__name__ in pks_to_delete.keys():
-                pks = pks_to_delete[klass.__name__]
-                objects_to_delete = (
-                    session.query(klass)
-                    .filter(getattr(klass, self.pk_names[klass.__name__]).in_(pks))
-                    .all()
-                )
-                deleted += len(objects_to_delete)
-                for obj in objects_to_delete:
-                    obj.deleted = True
-        return deleted
-
-
-class OSMLoader:
+class OSMLoader(BaseLoader):
+    METADATA_TABLE_NAME = "osm_metadata"
     POINT_TABLE_NAME = "osm_pisteet"
     # TODO: support linestrings later if needed
     # LINE_TABLE_NAME = "osm_viivat"
     POLYGON_TABLE_NAME = "osm_alueet"
-    HEADERS = {"User-Agent": "TARMO - Tampere Mobilemap"}
-    DATETIME_FORMAT = "%Y-%m-%d %H:%M:%S.%f"
 
     api_url = "https://overpass-api.de/api/interpreter"
     default_point = Point(0, 0)
@@ -138,68 +21,25 @@ class OSMLoader:
         connection_string: str,
         tags_to_include: Optional[dict] = None,
         tags_to_exclude: Optional[dict] = None,
-        overpass_api_url: Optional[str] = None,
-        point_of_interest: Point = default_point,
-        point_radius: float = 1,
+        **kwargs,
     ) -> None:
+        super().__init__(connection_string, **kwargs)
 
-        engine = create_engine(connection_string)
+        self.tags_to_include = (
+            tags_to_include if tags_to_include else self.metadata_row.tags_to_include
+        )
+        self.tags_to_exclude = (
+            tags_to_exclude if tags_to_exclude else self.metadata_row.tags_to_exclude
+        )
 
-        KoosteBase.prepare(engine, reflect=True)
-
-        self.Session = sessionmaker(bind=engine)
-
-        self.point_of_interest = point_of_interest
-        self.point_radius = point_radius
-
-        if overpass_api_url is not None:
-            self.api_url = overpass_api_url
-
-        with self.Session() as session:
-            self.syncher = Syncher(KoosteBase, session)
-
-            # OSM tables have their own metadata table, as they have extra tag filtering
-            metadata_row = session.query(KoosteBase.classes.osm_metadata).first()
-            self.last_modified = metadata_row.last_modified
-            self.tags_to_include = (
-                tags_to_include if tags_to_include else metadata_row.tags_to_include
-            )
-            self.tags_to_exclude = (
-                tags_to_exclude if tags_to_exclude else metadata_row.tags_to_exclude
-            )
-
-    def get_osm_objects(self) -> list:
+    def get_features(self) -> list:  # type: ignore[override]
         query = self.get_overpass_query()
         r = requests.post(self.api_url, headers=self.HEADERS, data=query)
         r.raise_for_status()
         data = r.json()
         return data["elements"]
 
-    def get_polygon_ring(self, ways: List) -> list:
-        """
-        Returns polygon ring, or empty list if the way is too short or not closed.
-
-        Multiple ways are concatenated together like OSM likes to do.
-        """
-        coordinate_lists = [way["geometry"] for way in ways]
-        coordinate_list = [node for node_list in coordinate_lists for node in node_list]
-        if len(coordinate_list) < 3:
-            return []
-        if (
-            coordinate_list[-1]["lat"] != coordinate_list[0]["lat"]
-            or coordinate_list[-1]["lon"] != coordinate_list[0]["lon"]
-        ):
-            # TODO: support ways that are not polygons
-            return []
-        return [
-            (
-                node["lon"],
-                node["lat"],
-            )
-            for node in coordinate_list
-        ]
-
-    def get_osm_feature(self, element: Dict[str, Any]) -> Optional[dict]:
+    def get_feature(self, element: Dict[str, Any]) -> Optional[dict]:  # type: ignore[override]  # noqa
         osm_id = element["id"]
         type = element["type"]
         tags = element["tags"]
@@ -243,28 +83,16 @@ class OSMLoader:
         }
         return flattened
 
-    def save_osm_feature(self, osm_object: Dict[str, Any], session: Session) -> bool:
-        table_cls = getattr(KoosteBase.classes, osm_object["table"])
-        new_obj = self.create_feature_for_object(table_cls, osm_object)
-        self.syncher.mark(new_obj)
-        try:
-            session.merge(new_obj)
-        except SQLAlchemyError:
-            LOGGER.exception(f"Error occurred while saving feature {osm_object}")
-        return True
-
-    def save_timestamp(self, session: Session) -> None:
-        metadata_row = session.query(KoosteBase.classes.osm_metadata).first()
-        metadata_row.last_modified = datetime.datetime.now()
-        session.merge(metadata_row)
-
     def get_overpass_query(self) -> str:
         """
         Construct the overpass query. This is a bit tricky if there are multiple tags
         we want.
         """
         # Of course, OSM uses lat,lon while we use lon,lat
-        around_string = f"around:{1000*self.point_radius},{self.point_of_interest.y},{self.point_of_interest.x}"  # noqa
+        if self.point_radius and self.point_of_interest:
+            around_string = f"around:{1000*self.point_radius},{self.point_of_interest.y},{self.point_of_interest.x}"  # noqa
+        else:
+            raise Exception("OSM queries require point and radius.")
         include_rows = (
             f"nwr[{tag}~\"^{'$|^'.join(values)}$\"]({around_string});"
             for tag, values in self.tags_to_include.items()
@@ -291,77 +119,31 @@ class OSMLoader:
 
         return query
 
-    def create_feature_for_object(
-        self, table_cls: Type[KoosteBase], osm_object: Dict[str, Any]  # type: ignore # noqa E501
-    ) -> KoosteBase:  # type: ignore
-        column_keys = set(table_cls.__table__.columns.keys())  # type: ignore
-        vals = {
-            key: osm_object[key]
-            for key in set(osm_object.keys()).intersection(column_keys)
-        }
-        vals["geom"] = "SRID=4326;" + vals["geom"]
-        return table_cls(**vals)  # type: ignore
+    def get_polygon_ring(self, ways: List) -> list:
+        """
+        Returns polygon ring, or empty list if the way is too short or not closed.
 
-    def save_features(
-        self, osm_objects: list, do_not_update_timestamp: Optional[bool] = False
-    ) -> str:
-        successful_actions = 0
-        with self.Session() as session:
-            for i, element in enumerate(osm_objects):
-                if i % 100 == 0:
-                    LOGGER.info(
-                        f"{100 * float(i) / len(osm_objects)}% - {i}/{len(osm_objects)}"
-                    )
-                osm_feature = self.get_osm_feature(element)
-                if osm_feature is not None:
-                    succeeded = self.save_osm_feature(osm_feature, session)
-                    if succeeded:
-                        successful_actions += 1
-                else:
-                    LOGGER.debug(f"OSM feature {osm_feature} empty")
-
-            if not do_not_update_timestamp:
-                self.save_timestamp(session)
-            deleted_items = self.syncher.finish(session)
-            session.commit()
-        msg = f"{successful_actions} inserted or updated. {deleted_items} deleted."
-        LOGGER.info(msg)
-        return msg
+        Multiple ways are concatenated together like OSM likes to do.
+        """
+        coordinate_lists = [way["geometry"] for way in ways]
+        coordinate_list = [node for node_list in coordinate_lists for node in node_list]
+        if len(coordinate_list) < 3:
+            return []
+        if (
+            coordinate_list[-1]["lat"] != coordinate_list[0]["lat"]
+            or coordinate_list[-1]["lon"] != coordinate_list[0]["lon"]
+        ):
+            # TODO: support ways that are not polygons
+            return []
+        return [
+            (
+                node["lon"],
+                node["lat"],
+            )
+            for node in coordinate_list
+        ]
 
 
 def handler(event: Event, _) -> Response:
     """Handler which is called when accessing the endpoint."""
-    response: Response = {"statusCode": 200, "body": json.dumps("")}
-    db_helper = DatabaseHelper()
-    try:
-        params = {}
-        point = (
-            Point(event["close_to_lon"], event["close_to_lat"])
-            if "close_to_lon" in event and "close_to_lat" in event
-            else None
-        )
-        if point:
-            params["point_of_interest"] = point
-        radius = event.get("radius", None)
-        if radius:
-            params["point_radius"] = radius
-
-        loader = OSMLoader(db_helper.get_connection_string(), **params)
-        osm_objects = loader.get_osm_objects()
-
-        msg = loader.save_features(
-            osm_objects, event.get("do_not_update_timestamp", False)
-        )
-        LOGGER.info(msg)
-        response["body"] = json.dumps(msg)
-
-    except Exception:
-        # LOGGER.exception("Uncaught error occurred")
-        response["statusCode"] = 500
-
-        exc_info = sys.exc_info()
-        exc_string = "".join(traceback.format_exception(*exc_info))
-        response["body"] = exc_string
-        LOGGER.exception(exc_string)
-
-    return response
+    return base_handler(event, OSMLoader)
