+++ backend/test/test_services.py
@@ -73,6 +73,7 @@ def populate_two_pages_of_lipas(create_db, main_db_params, lipas_loader_url):
     }
     r = requests.post(lipas_loader_url, data=json.dumps(payload))
     data = r.json()
+    print(data)
     assert data["statusCode"] == 200, data["body"]
